feat: ui-dashboard-stream-timeline-split-004 - Compose dashboard shell without behavior regression
